Adjusted get_makeover_groups code
The get_makeover_groups() function has been moved from appearance.py to options.py in an attempt to resolve a circular import when options.py is initialized first. Considering the code had no use other than to add options to the GUI, it seems appropriate for options.py.

The code in options.py that populates the makeover_groups was previously at the module level and would run during the initialization of options.py. That code relied on a method in appearance.py, but appearance.py imports options.py. So options.py tries to load appearance.py, which turns around and tries to load options.py, resulting in a deadlock.

With the code now within a function, options.py will no longer attempt to import from appearance.py during the middle of initialization, eliminating the circular import. The downside is that makeover_groups will remain unpopulated until a call to get_makeover_groups() has been made. A call to that function has been added to beyondchaos.py.

Removed debug print from remonsterate.py that I accidentally left in.
